Refactor: rename files inside components, grouping & move folders
