+++ client-mvp-04/src/components/common/footer/footer.scss
@@ -1,4 +1,4 @@
-@import "../../sass/_main.scss";
+@import "../../../sass/main";
 
 .app-footer {
   display: flex;
