+++ client-mvp-04/src/components/common/link/link.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import RedirectLink from './index';
+import RedirectLink from './link';
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
