+++ client-mvp-04/src/components/error/error.scss
