--- client-mvp-04/src/reducers/authReducer.js
@@ -1,17 +1,0 @@
-import { LOGIN } from '../actions/types';
-
-const authDefaultState = {
-    loggedIn : false
-}
-
-export default (state = authDefaultState, {type, payload}) => {
-    switch (type) {
-        case LOGIN:
-            return {
-                ...state,
-                loggedIn : true
-            }
-        default:
-            return state;
-    }
-}
