--- client-mvp-04/src/store/configureStore.js
@@ -1,9 +1,0 @@
-import { createStore } from 'redux';
-import allReducers from '../reducers';
-
-export default () => {
-    return createStore(
-        allReducers,
-        window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
-    );
-}
