+++ client-mvp-04/src/store/store.js
@@ -1,0 +1,9 @@
+import { createStore } from 'redux';
+import allReducers from './reducers';
+
+export default () => {
+  return createStore(
+    allReducers,
+    window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  );
+};
